can/python: Implemented in graph execution as part of the GUI event loop.
This method is by far *the easiest* and *most responsive* of all the
presented implementations, all while being just as fast/faster.

Threaded examples, while interesting to look at, are more complex
with no real benefit.

`can_graph_gui_threaded_responsive.py` was reworked
    into `can_graph_gui_advanced.py`

`can_graph_gui_threaded_simple.py` was reworked
    into `can_graph_gui_simple.py`
